大二下/人工智能导论/情感分析/wzf2000/SentimentAnalysis/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
altgraph=0.17=pyhd3eb1b0_0
blas=1.0=mkl
bzip2=1.0.8=h1de35cc_0
ca-certificates=2021.1.19=hecd8cb5_1
certifi=2020.12.5=py38hecd8cb5_0
cycler=0.10.0=py38_0
ffmpeg=4.3=h0a44026_0
freetype=2.10.4=ha233b18_0
gettext=0.19.8.1=hb0f4f8b_2
gmp=6.2.1=h23ab428_2
gnutls=3.6.5=h91ad68e_1002
intel-openmp=2019.4=233
jieba=0.42.1=pypi_0
joblib=1.0.1=pypi_0
jpeg=9b=he5867d9_2
kiwisolver=1.3.1=py38h23ab428_0
lame=3.100=h1de35cc_0
lcms2=2.11=h92f6f08_0
libcxx=10.0.0=1
libedit=3.1.20191231=h1de35cc_1
libffi=3.3=hb1e8313_2
libiconv=1.16=h1de35cc_0
libpng=1.6.37=ha441bb4_0
libtiff=4.1.0=hcb84e12_1
libuv=1.40.0=haf1e3a3_0
lz4-c=1.9.3=h23ab428_0
macholib=1.14=pyhd3eb1b0_1
matplotlib=3.3.4=py38hecd8cb5_0
matplotlib-base=3.3.4=py38h8b3ea08_0
mkl=2019.4=233
mkl-service=2.3.0=py38h9ed2024_0
mkl_fft=1.3.0=py38ha059aab_0
mkl_random=1.1.1=py38h959d312_0
ncurses=6.2=h0a44026_1
nettle=3.4.1=h3018a27_0
ninja=1.10.2=py38hf7b0b51_0
numpy=1.20.3=pypi_0
olefile=0.46=py_0
openh264=2.1.0=hd9629dc_0
openssl=1.1.1j=h9ed2024_0
pandas=1.2.3=py38hb2f4e1b_0
pillow=8.1.2=py38h5270095_0
pip=21.0.1=py38hecd8cb5_0
pycryptodome=3.10.1=py38h659dea6_0
pyinstaller=3.6=py38ha441bb4_5
pyparsing=2.4.7=pyhd3eb1b0_0
pypinyin=0.41.0=pypi_0
python=3.8.8=h88f2d9e_4
python-dateutil=2.8.1=pyhd3eb1b0_0
pytorch=1.8.0=py3.8_0
pytz=2021.1=pyhd3eb1b0_0
readline=8.1=h9ed2024_0
scikit-learn=0.24.2=pypi_0
scipy=1.6.3=pypi_0
setuptools=52.0.0=py38hecd8cb5_0
six=1.15.0=py38hecd8cb5_0
sklearn=0.0=pypi_0
sqlite=3.33.0=hffcf06c_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=hb0a8c7a_0
torchaudio=0.8.0=py38
torchvision=0.9.0=py38_cpu
tornado=6.1=py38h9ed2024_0
tqdm=4.59.0=pypi_0
typing_extensions=3.7.4.3=pyha847dfd_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.5=h41d2c2f_0
